Update README.md
									Nick Johnson
This report is to analyze the finger force in the first and third phalanx.  The measurement collects the force in extending the first and third phalanx. It also collects the force in flexion the third phalanx.  The phalanx refers to the three bones in each finger.  The first phalanx is the biggest also known as the distal phalanx and the third phalanx is the smallest bone which is referred to as the proximal phalanx.  Thumbs were not included in this research.  There were 15 subjects that used each finger (values I for index, M for middle, R for ring and L for little) 3 times in 3 different locations.  Breaks were taken in between measurements to prevent muscle fatigue to ensure accurate results.  In the research it doesn’t state whether the hand used was right, left, dominate, or weak.  This could hinder accurate measurements when comparing strengths among subjects.   
The first model is a two ANOVA table.  Looking at the graphs in Fig. 3 we can see some subject have all positive or negative residuals.  This would suggest an effect that has not been considered in the testing.  This would also suggest that the variances in all groups aren’t equal.  In the second step the mixed effect model is used.  Here the assumption of random effects are independent from each other.  Residuals now have significant change from one position to the next.  Though the model has a better fit, the graphs denote residual variability issues from locations.  Now modeling the variance structures to test model residuals is used.  When comparing the boxplots for mixed effect and modeling variance structures there is less variance between locations.  The residuals for fingers don’t really change from finger to finger as proportionally they stay the same.  But the index finger does look vastly different from the rest of the fingers.  Lastly, modeling the correlation matrix is used to compare to the variance matrix.  The boxplots for location and fingers look like a more closely fitted variance than any other model.  As we look at the AIC/BIC from one model to the next they decrease in size.  The AIC started at 3062.614 and shrunk to 2163.984 with the correlation matrix model.  The BIC started at 3122.696 and ended up at 2292.731 with the correlation matrix model as well.  
There are concerns to the data.  Are there enough people to accurately measure these metrics?  Why are only 2 areas of the finger tested and not all 3?  Lastly, why aren’t the tests performed at each location?  I think the test has some missing data to be completely thorough.  The model doesn’t seem to fit a linear model without other factors being considered.